Completing refactor
