+++ src/components/AppContainer.js
@@ -2,12 +2,12 @@ import 'react-select/dist/react-select.css';
 import { connect } from 'react-redux';
 import { routerActions } from 'react-router-redux';
 import { TYPES as PICK_TYPES, clearNodes, clearObjects } from '../modules/picklist/picklistReducer';
-import { loadBuildinfo } from '../reducers/auth';
+import { loadBuildinfo } from './appReducer';
 import { clearSearch } from '../modules/objectsearch/actions';
 import App from './AppComponent';
 import Notifyable from './Notifyable';
 import {clearStats} from '../modules/storagefacility/reducers/stats';
-import {setMuseumId, setCollectionId } from '../reducers/auth';
+import {setMuseumId, setCollectionId } from './appReducer';
 import {clearRoot, loadRoot} from '../modules/storagefacility/reducers/grid/nodes';
 import MuseumId from '../shared/models/museumId';
