+++ src/components/Authenticated.js
@@ -3,7 +3,7 @@ import Loader from 'react-loader';
 import App from './AppContainer';
 import Login from '../modules/login/LoginContainer';
 import {connect} from 'react-redux';
-import { setUser, loadActor } from '../reducers/auth';
+import { setUser, loadActor } from './appReducer';
 import fakeUserInfo from '../../fake_security.json';
 import jwtDecode from 'jwt-decode';
 import { emitError } from '../shared/errors/emitter';
